+++ src/components/dialogs/ConfirmTermsDialog.tsx
@@ -33,7 +33,8 @@ interface IntConfirmOffer {
 interface IntDialogProps {
   title?: string;
   open: boolean;
-  handleButtonEvent?: (type: string) => void;
+  handleConfirm?: () => void;
+  handleClose?: (state: boolean) => void;
   isProgress?: boolean;
   offerObj?: IntConfirmOffer;
 }
@@ -41,17 +42,14 @@ interface IntDialogProps {
 const ConfirmTermsDialog: React.FC<IntDialogProps> = ({
   title = 'Confirm',
   open = false,
-  handleButtonEvent,
+  handleConfirm,
+  handleClose,
   isProgress = false,
   children,
   offerObj,
 }) => {
   const [isAgreed, setIsAgreed] = useState(false);
   const { t } = useTranslation();
-
-  const handleButton = (type: string) => {
-    handleButtonEvent(type);
-  };
 
   function splitWithFirstOcc(str: string) {
     const regX = /:(.*)/s;
@@ -60,7 +58,7 @@ const ConfirmTermsDialog: React.FC<IntDialogProps> = ({
 
   return (
     <Dialog open={open}>
-      <DialogHeader closeWithIcon onCloseWithIcon={() => handleButton('close')} title={title} />
+      <DialogHeader closeWithIcon onCloseWithIcon={() => handleClose(false)} title={title} />
       <DialogContent dividers sx={{ py: 3 }}>
         {children ? (
           children
@@ -68,23 +66,23 @@ const ConfirmTermsDialog: React.FC<IntDialogProps> = ({
           <>
             <Box sx={{ mb: 1 }}>
               {offerObj?.offerCount !== 0 && (
-                <p>
+                <Box>
                   <Trans i18nKey={'dialog.offerDetails.confirmTermsTitle'} count={offerObj.offerCount} />
-                </p>
+                </Box>
               )}
-              <p>
+              <Box>
                 {t('dialog.offerDetails.cofirmTermsSubtitle')}
                 {offerObj ? (
-                  <strong style={{ margin: '0 5px' }}>{`${splitWithFirstOcc(offerObj.provider)[0]}.` || '-.'}</strong>
+                  <b style={{ margin: '0 5px' }}>{`${splitWithFirstOcc(offerObj.provider)[0]}.` || '-.'}</b>
                 ) : (
                   '-.'
                 )}
-              </p>
-              <p>{t('dialog.offerDetails.confirmHeading')}</p>
+              </Box>
+              <Box>{t('dialog.offerDetails.confirmHeading')}</Box>
             </Box>
-            <p>(1) {t('dialog.offerDetails.point1')}</p>
-            <p>(2) {t('dialog.offerDetails.point2')}</p>
-            <p>(3) {t('dialog.offerDetails.point3')}</p>
+            <Box>(1) {t('dialog.offerDetails.point1')}</Box>
+            <Box>(2) {t('dialog.offerDetails.point2')}</Box>
+            <Box>(3) {t('dialog.offerDetails.point3')}</Box>
             <FormControlLabel
               control={<Checkbox checked={isAgreed} onChange={() => setIsAgreed(!isAgreed)} name="gilad" />}
               label={t('content.common.agree')}
@@ -93,7 +91,7 @@ const ConfirmTermsDialog: React.FC<IntDialogProps> = ({
         )}
       </DialogContent>
       <DialogActions>
-        <Button variant="outlined" disabled={isProgress} onClick={() => handleButton('close')}>
+        <Button variant="outlined" disabled={isProgress} onClick={() => handleClose(false)}>
           {t('button.cancel')}
         </Button>
         <LoadingButton
@@ -102,7 +100,7 @@ const ConfirmTermsDialog: React.FC<IntDialogProps> = ({
           disabled={isProgress || !isAgreed}
           label={t('button.confirm')}
           loadIndicator={t('content.common.loading')}
-          onButtonClick={() => handleButton('confirm')}
+          onButtonClick={() => handleConfirm()}
           loading={isProgress}
           sx={{ ml: 3 }}
         />