+++ src/components/policies/UsagePolicyItem.tsx
@@ -20,11 +20,12 @@
 
 import { Box, FormControl, FormControlLabel, FormLabel, Radio, RadioGroup, Stack } from '@mui/material';
 import { Input, SelectList } from 'cx-portal-shared-components';
+import { isEmpty } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
 import { setDurationUnit, setPurposeValue } from '../../features/provider/policies/slice';
 import { useAppDispatch, useAppSelector } from '../../features/store';
-import { DURATION_UNITS, PURPOSE_VALUES } from '../../utils/constants';
+import { DURATION_UNITS, ONLY_NUM_REGEX, PURPOSE_VALUES } from '../../utils/constants';
 
 interface FreeTextProps {
   restrictionType: string;
@@ -51,8 +52,10 @@ export default function UsagePolicyItem({
 
   function checkErrors() {
     if (constraintType === t('content.policies.duration')) {
-      const result = /^[1-9]\d*$/.test(inputFreeText);
+      const result = ONLY_NUM_REGEX.test(inputFreeText);
       return !result;
+    } else if (constraintType === t('content.policies.purpose')) {
+      return isEmpty(purposeValue);
     } else if (restrictionType === 'RESTRICTED' && inputFreeText === '') {
       return true;
     }
@@ -83,7 +86,9 @@ export default function UsagePolicyItem({
                   required
                   error={checkErrors()}
                   onChange={e => {
-                    setInputFreeText(e.target.value);
+                    const val = e.target.value;
+                    const check = ONLY_NUM_REGEX.test(val) || val === '';
+                    if (check) setInputFreeText(val);
                   }}
                   sx={{ minWidth: 250 }}
                 />
@@ -92,14 +97,13 @@ export default function UsagePolicyItem({
                 <FormControl sx={{ minWidth: 150 }} size="small">
                   <SelectList
                     keyTitle="title"
-                    value={durationUnit}
-                    defaultValue={DURATION_UNITS[0]}
+                    defaultValue={durationUnit}
                     items={DURATION_UNITS}
                     label={t('content.policies.selectDuration')}
                     placeholder={t('content.policies.selectDuration')}
                     disableClearable={true}
                     onChangeItem={e => {
-                      dispatch(setDurationUnit(e.value));
+                      dispatch(setDurationUnit(e));
                     }}
                   />
                 </FormControl>
@@ -108,7 +112,6 @@ export default function UsagePolicyItem({
                 <FormControl sx={{ minWidth: 250 }} size="small">
                   <SelectList
                     keyTitle="title"
-                    value={purposeValue}
                     defaultValue={purposeValue}
                     items={PURPOSE_VALUES}
                     label={t('content.policies.purposeLabel')}